# apps/backend/openapi.test3.yaml
openapi: 3.0.3
info:
  title: X-Ear CRM API (auto-generated)
  version: 0.0.0-auto
servers:
- url: http://localhost:5003
paths:
...
  /api/patients/{patient_id}/ereceipts:
    get:
      summary: GET /api/patients/{patient_id}/ereceipts
      operationId: patient_subresources_get_patient_ereceipts
      tags:
      - Patients
      parameters:
      - name: patient_id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Success
    post:
      summary: POST /api/patients/{patient_id}/ereceipts
      operationId: patient_subresources_create_patient_ereceipt
      tags:
      - Patients
      parameters:
      - name: patient_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Success
  /api/patients/{patient_id}/ereceipts/{ereceipt_id}:
    put:
      summary: PUT /api/patients/{patient_id}/ereceipts/{ereceipt_id}
      operationId: patient_subresources_update_patient_ereceipt
      tags:
      - Patients
      parameters:
      - name: patient_id
        in: path
        required: true
        schema:
          type: string
      - name: ereceipt_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Success
    delete:
      summary: DELETE /api/patients/{patient_id}/ereceipts/{ereceipt_id}
      operationId: patient_subresources_delete_patient_ereceipt
      tags:
      - Patients
      parameters:
      - name: patient_id
        in: path
        required: true
        schema:
          type: string
      - name: ereceipt_id
        in: path
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Success
  /api/patients/{patient_id}/hearing-tests:
    get:
      summary: GET /api/patients/{patient_id}/hearing-tests
      operationId: patient_subresources_get_patient_hearing_tests
      tags:
      - Patients
      parameters:
      - name: patient_id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Success
    post:
      summary: POST /api/patients/{patient_id}/hearing-tests
      operationId: patient_subresources_add_patient_hearing_test
      tags:
      - Patients
      parameters:
      - name: patient_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Success
  /api/patients/{patient_id}/hearing-tests/{test_id}:
    put:
      summary: PUT /api/patients/{patient_id}/hearing-tests/{test_id}
      operationId: patient_subresources_update_patient_hearing_test
      tags:
      - Patients
      parameters:
      - name: patient_id
        in: path
        required: true
        schema:
          type: string
      - name: test_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Success
    delete:
      summary: DELETE /api/patients/{patient_id}/hearing-tests/{test_id}
      operationId: patient_subresources_delete_patient_hearing_test
      tags:
      - Patients
      parameters:
      - name: patient_id
        in: path
        required: true
        schema:
          type: string
      - name: test_id
        in: path
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Success
...
  /api/patients/{patient_id}/notes:
    get:
      summary: GET /api/patients/{patient_id}/notes
      operationId: patient_subresources_get_patient_notes
      tags:
      - Patients
      parameters:
      - name: patient_id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Success
    post:
      summary: POST /api/patients/{patient_id}/notes
      operationId: patient_subresources_create_patient_note
      tags:
      - Patients
      parameters:
      - name: patient_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Success
  /api/patients/{patient_id}/notes/{note_id}:
    put:
      summary: PUT /api/patients/{patient_id}/notes/{note_id}
      operationId: patient_subresources_update_patient_note
      tags:
      - Patients
      parameters:
      - name: patient_id
        in: path
        required: true
        schema:
          type: string
      - name: note_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Success
    delete:
      summary: DELETE /api/patients/{patient_id}/notes/{note_id}
      operationId: patient_subresources_delete_patient_note
      tags:
      - Patients
      parameters:
      - name: patient_id
        in: path
        required: true
        schema:
          type: string
      - name: note_id
        in: path
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Success
...